Add 3rd party asset loading methods
