+++ projects/plugins/jetpack/extensions/blocks/map/block-editor-assets.json
